Drop support for OpenSearch 1.x
