+++ .github/workflows/test.yml
@@ -44,9 +44,6 @@ jobs:
           - database: sqlite
             backend: elasticsearch9
           - database: sqlite
-            backend: opensearch1
-            opensearch-version: 1.3.0
-          - database: sqlite
             backend: opensearch2
             opensearch-version: 2.10.0
           - database: sqlite
@@ -182,8 +179,6 @@ jobs:
             pip install "elasticsearch>=8,<9"
           elif [[ "${{ matrix.backend }}" == "elasticsearch9" ]]; then
             pip install "elasticsearch>=9,<10"
-          elif [[ "${{ matrix.backend }}" == "opensearch1" ]]; then
-            pip install "opensearch-py>=1,<2" requests
           elif [[ "${{ matrix.backend }}" == "opensearch2" ]]; then
             pip install "opensearch-py>=2,<3" requests
           elif [[ "${{ matrix.backend }}" == "opensearch3" ]]; then
