+++ modelsearch/test/settings.py
@@ -99,7 +99,6 @@ MODELSEARCH_BACKENDS = {
             "elasticsearch7": "modelsearch.backends.elasticsearch7",
             "elasticsearch8": "modelsearch.backends.elasticsearch8",
             "elasticsearch9": "modelsearch.backends.elasticsearch9",
-            "opensearch1": "modelsearch.backends.opensearch1",
             "opensearch2": "modelsearch.backends.opensearch2",
             "opensearch3": "modelsearch.backends.opensearch3",
         }[SEARCH_BACKEND]
@@ -114,7 +113,7 @@ if SEARCH_BACKEND in ["elasticsearch7", "elasticsearch8", "elasticsearch9"]:
         "ca_certs": os.environ.get("ELASTICSEARCH_CA_CERTS"),
     }
 
-if SEARCH_BACKEND in ["opensearch1", "opensearch2", "opensearch3"]:
+if SEARCH_BACKEND in ["opensearch2", "opensearch3"]:
     OPENSEARCH_URL = os.getenv("OPENSEARCH_URL", "http://localhost:9200")
     MODELSEARCH_BACKENDS["default"]["URLS"] = [OPENSEARCH_URL]
     MODELSEARCH_BACKENDS["default"]["INDEX_PREFIX"] = "modelsearchtest_"
