--- modelsearch/tests/test_opensearch1_backend.py
@@ -1,7 +1,0 @@
-from django.test import TestCase
-
-from .elasticsearch_common_tests import ElasticsearchCommonSearchBackendTests
-
-
-class TestOpenSearch1SearchBackend(ElasticsearchCommonSearchBackendTests, TestCase):
-    backend_path = "modelsearch.backends.opensearch1"
